ethclient: move test to 'cg' namespaced file, allows testing @ethereum/go-ethereum
Date: 2021-03-15 13:35:41-05:00
